third_party/WebKit/LayoutTests/accessibility/aria-setsize-posinset-expected.txt

Radio 1

This tests that aria-posinset and aria-setsize are exposed to accessibility correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS list.childAtIndex(0).setSize is 10
PASS list.childAtIndex(1).setSize is 10
PASS list.childAtIndex(0).posInSet is 3
PASS list.childAtIndex(1).posInSet is 5
PASS listbox.childAtIndex(0).setSize is 4
PASS listbox.childAtIndex(1).setSize is 4
PASS listbox.childAtIndex(2).setSize is 4
PASS listbox.childAtIndex(3).setSize is 4
PASS listbox.childAtIndex(0).posInSet is 1
PASS listbox.childAtIndex(1).posInSet is 2
PASS listbox.childAtIndex(2).posInSet is 3
PASS listbox.childAtIndex(3).posInSet is 4
PASS radio1.setSize is 1
PASS radio1.posInSet is 1
PASS successfullyParsed is true

TEST COMPLETE
